+++ CHANGELOG.md
@@ -51,6 +51,7 @@
 * st-modal2: renamed to sds-modal
 * st-color-picker: Removed component
 * st-two-list-selection: Remove component
+* st-dynamic-table: Removed component
 
 **Deprecated:**